[import-w3c-tests] Provide useful errors with invalid paths https://bugs.webkit.org/show_bug.cgi?id=274880 rdar://128981568
Reviewed by NOBODY (OOPS!).

This handles all three notable cases:

 * a path not beginning with `web-platform-tests`,
 * a path not existing,
 * a path referring to an individual file.

While we previously had the first one, this moves it to where we
previously discounted any such path rather than having a separate
check.

Then, just before we start iterating through all directories, we now
validate the path actually exists.

As a drive-by, this improves logging when we're ensuring the path
specified is a WPT repo, and fixes the tests after some earlier
regression.