Add a Test-Sequence header to the webkitpy <-> WKTR/DRT server protocol https://bugs.webkit.org/show_bug.cgi?id=298004
Reviewed by NOBODY (OOPS!).

This keeps count of the number of run test commands sent on both sides
of the protocol, allowing some basic synchronization.

It is less ideal than setting some abstract UUID (because
incrementing accidentally is possible), but this is landable in a much
lower-impact way that doesn't break the communication protocol,
allowing older WKTR/DRT builds (e.g., from build archives) to still
work with current webkitpy.

We report failures as CRASH, because a driver synchronization issue is
a significant one. Of course, by the point this occurs, it's reporting
a _following_ test as CRASH (and almost certainly the _immediately_
following test), and not the one that caused the synchronization issue
to start with.

Explanation of why this fixes the bug (OOPS!).